pyembed: Fix types for linux/aarch64
The c_char and wchar_t types on linux/aarch64 are unsigned, so some
casts to "*const i8" were turned to "*const c_char" and the type
definition for wchar_t to be used from libc.
libc was a transient dependency (from pyo3 at least) and is now direct.

Also fixes default_target to return the correct value on linux/aarch64.